Add files via upload
# Contact Management System

 Objective
The objective of this project is to implement a simple system to manage contacts. This C++ program allows users to add, view, and delete contacts, each of which includes a name and a phone number.

## Features
Add Contact:  Add a new contact by providing a name and phone number.
View Contacts:  Display the list of all contacts.
Delete Contact: Remove a contact by specifying its name.
Menu-Driven Interface: User-friendly interface for easy navigation and interaction.